# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API en Laravel",
        "description": "Documentación de la API con Swagger en Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener lista de productos",
                "operationId": "496dc08306524d77f18d124d1d6838aa",
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida correctamente"
                    },
                    "401": {
                        "description": "No autorizado - Token ausente o inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No autenticado o token inválido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "description": "Crea un producto nuevo. Requiere autenticación con token JWT.",
                "operationId": "48b59f126c849beb4d4de4ff83b7945f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Laptop Gamer"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Laptop con 16GB RAM y RTX 4060"
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500.5
                                    },
                                    "peso": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.5
                                    },
                                    "disponible": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "fecha_vencimiento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    },
                                    "publicado_en": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-09-30T12:30:00Z"
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado correctamente"
                    },
                    "401": {
                        "description": "No autorizado - Token ausente o inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "No autenticado o token inválido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    },
                    "500": {
                        "description": "Error interno"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Usa un token JWT para autenticar",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}